The application follows the MVC pattern, a separate view has been created, the view is connected to the controller, a new model has been created to count the number of sessions, when the stop button is pressed, old models are saved and new ones are created (For convenience of working with Realm in the future).
