+++ Pomodoro/Support/Assets.xcassets/AppIcon.appiconset/Contents.json
